requirements.txt
APScheduler==3.10.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.9.0
bs4==0.0.1
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
chromedriver-binary==132.0.6794.2.0
click==8.1.7
exceptiongroup==1.1.1
Flask==3.1.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==24.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==5.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.9.0
six==1.16.0
slack_bolt==1.21.2
slack_sdk==3.33.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.19
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
